requirements.txt
mss==10.0.0
pywin32==306
pynput==1.7.6
httpx==0.27.2
pillow==10.4.0
python-dotenv==1.0.1
SpeechRecognition==3.10.4
PyAudio==0.2.14
sarvamai==0.1.19

mss==10.0.0
pywin32==306
pynput==1.7.6
httpx==0.27.2
pillow==10.4.0
python-dotenv==1.0.1

